+++ test/src/string_test.dart
@@ -1,3 +1,11 @@
+//
+//  handy_extensions
+//  string_test
+//
+//  Created by Ngonidzashe Mangudya on 15/02/2024.
+//  Copyright (c) 2024 ModestNerds, Co
+//
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:handy_extensions/handy_extensions.dart';